addressed review comments
