+++ rapids-4-spark-tools/.gitignore
@@ -1,0 +1,4 @@
+.idea
+*.iml
+target
+event_log_profiling.log
